requirements.txt
Flask
flask-cors
Pillow
pytesseract
opencv-python
numpy